:disappointed: hardware: UltimateHackingKeyboard is an honorable mention
Prokura Innovations pointed out to me the licenses used on the Ultimate
Hacking Keyboard. Six months ago from today, UHB changed the licenses on
all of their repositories. I'm not sure what was used before, but now it
is a custom proprietary license that restricts one of the Four Freedoms
of Free Software.

So, I made this clear and distinguished it from the other case studies
on the page. It is too bad it is proprietary!

Signed-off-by: Justin W. Flory <git@jwf.io>