:new: :microphone: foss: Add Needs Assessment Template
This is a template we used in LibreCorps to interview teams in the
initial on-boarding interviews. Given that I am writing a needs
assessment survey for incoming teams, I think this is a useful document
to surface. I will probably iterate further on this for the next cohort.

See librecorps/resources@42e50b23fc1e104611f782dc41e9d367d3f17035.

Signed-off-by: Justin W. Flory <git@jwf.io>