:new: :space_invader: Communities: Initial commit, Public Lab
First commit with an introduction to the Public Lab community. I used
the quote from the Public Labs website as pointed out by Jeanette Lim.
The Public Lab entry is part of a new section on "Communities". I hope
to add more communities of practice for our teams to engage with here
over time.

cc: @lizbarry

Signed-off-by: Justin W. Flory <git@jwf.io>