:new: meta: Initial commit
This commit sets up a new Heuristic category, called "Meta". This is
going to be a place where I store all of my templates, notes, and other
data that I probably will _not_ share with teams, but I will refer to
often to help guide me with Open Source Mentorship.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>