:microphone: meta: Import and update needs assessment
For the past month, I have referred back to this Needs Assessment quite
a bit! However, I have ad-libbed it a lot. So, this commit gets it up to
speed with reality, based on what I _actually_ ask these days on the
Open Source induction calls.

Additionally, I also moved this document from the "FOSS" category to the
"Meta" category, since this document is really for me and not the teams
I am providing mentorship to.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>