:new: :mag: documentation: Project Hall of Fame, initial commit
This commit starts a new page to aggregate examples of good Open Source
documentation projects. It is based on an initial crawl of projects that
I was already aware of, combined with what appeared on the first page of
Google. So, hopefully this is a good start to good examples of Open
Source docs. I anticipate adding more examples to this over time, as I
find them!

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>