:new: :memo: foss: Add business sustainability / training services doc
This commit creates a new page based on conversations from October 2020
check-in calls with the Open Source Mentorship programme at the UNICEF
Innovation Fund. This adds a new page generally about business
sustainability.

To seed the page, I added the requested part about using training as a
way to support your business. This uses the big example in the room, Red
Hat, but also looks to two Innovation Fund inductees.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>