:mag: :bulb: foss: Add GPL comparison page
This addition is one of my favorites and required a good deal of
reading. This is my overall assessment on the differences of the GPLv2
to the GPLv3. I make a general assessment and overview of the
differences. I also try to represent the diversity of opinions around
the updated license.

Special shout-out to @deejoe for talking through it in `#rit-foss` some
weeks ago.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>