:new: :robot: hardware: Add a general introduction page
This commit creates a general overview to what Open Hardware is, points
to existing examples of Open Source Hardware projects, and explains the
key benefits to choosing to go open with a hardware product.

This is in follow-up to feedback from the Comunicación Aumentativa team
from Chile on November 27th, 2020.

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>