:alembic: ci: Enable htmlproofer checks on new Pull Requests (#6)
Added in CircleCI-Public/hugo-orb#45 and CircleCI-Public/hugo-orb#46.

This commit turns HTML Proofer back on. Previously, it failed because of
unexpected error codes from sites linked elsewhere. This uses the new
feature flags added upstream in an effort to hopefully gain the helpful
feedback of when links in the site are broken.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>