:memo: Document phases/modules of Open Source Mentorship programme (#8)
* Simplify URL schema for mentor templates doc

The entire `meta/` directory mostly pertains to Open Source Mentorship.
This is why it is meta. No need to be so declarative in the use of
"mentorship" in files. Also adjusted file weight.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

* Pull out 12-month roadmap from Mentorship overview doc

This will live in a new document specifically about Open Source
Mentorship programme modules offered to inductees.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

* Describe three of five modules, structure programming options.

This commit describes three of the modules I offer to Innovation Fund
grantees in the existing Open Source Mentorship programme. The other two
are identified but not yet described.

There is also a template structure for two programming options: the
12-month roadmap originally included in the overview, and an
experimental 3-month program with mix-and-match modules.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

* Document remaining two modules

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

* Add module programming and expand on 12-month contract milestones.

This commit combines the information I drafted for the Open Source
workplans in this public documentation. It is mostly a direct copy+paste
with some context added in at different parts.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>