:new: :art: theme: Update git submodule to forked theme (closes #13)
This commit updates the UNICEF Open Source Inventory to use a new git
submodule for the Hugo theme. The new submodule comes from a forked
repository of the original Dot theme used in the project.

In the future clones and CI pipelines, it should use the new theme. This
allows us to test new changes made to the theme locally from this Hugo
site (without deploying a new version to production to test).

Closes #13.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>