:new: :satellite: Add contributing guidelines, better docs (closes #17) (#19)
This commit adds new contributing guidelines. The contributing
guidelines also provide an explanation of cloning the repository with
the necessary git submodule. This should avoid future headaches for
anyone who works with or contributes to this repository.

Also added a note on installing AsciiDoctor, thanks to @Idadelveloper
pointing out the undocumented dependency for the project.

Closes #17.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>