:package: :arrow_down: theme: Add OpenGraph defaults and more alt tags
This commit uses a new feature first proposed by @MuluhGodson, later
reviewed by @somratpro, to add OpenGraph metadata into the site. For
now, this is just a basic default with site description, authors, and an
image that we have used with other Open Source publications on
unicef.org.

I also added some alt tags to the social media icons, since this was
integrated upstream some time ago too.

Contributes to #18, but does not yet close it.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>