:new: :mexico: l10n: Add base Spanish translations for Inventory
This is a first attempt at integrating multilingual content into the
Open Source Inventory. Since the theme has good support for multiple
languages, it is helpful for us to try and leverage this.

For now, I am translating Spanish since I know the language, but there
is possibility to add others in the future too.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>