Make blockquotes more pretty and make citation more clear (#23)
* content: Add image and user-quote class to citation

This commit tackles issue #14 aimed at enhancing the citation look. In case
there is an image, the link is placed after the paragraph. The 'user-quote'
class is added so that unique styles perculiar to quotes from an individual
are added. Therefore, the blockquote is coming from an individual, the
'user-quote' class should be included.

Signed-off-by: Ida Delphine <idadelm@gmail.com>

* Add UNICEF Inventory Hugo theme as a git submodule

Signed-off-by: Ida Delphine <idadelm@gmail.com>

* themes: Add upstream changes

Signed-off-by: Ida Delphine <idadelm@gmail.com>