:new: :camera_flash: config: Change OpenGraph image to UNI309823
This commit changes the OpenGraph image to an image I found in the
UNICEF archives for other related Open Source communications from the
Office of Innovation at UNICEF HQ. I believe it is a better
representation of this repository's mission and goals than the original
photo. Even if I like both. :grinning:

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>