:wrench: Issue templates: Add a config file with more options
Disable blank issues because I strongly believe all content should use a
template. If a template is missing, it should be added. If an existing
template is odd or not working, it should be changed or removed.

I also added more outbound links to the UNICEF Innovation Fund Discourse
forum for extra measure.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>