:new: :mag: config: Add Google Analytics support
I am adding Google Analytics support because it is the only tool in my
personal toolbox for measuring site engagement, and it is already built
into the theme used for this Hugo site.

The hope is that this change will enable us to make insights into what
content is popular and how we can optimize or improve content based on
what people seem to be reading or most interested about.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>