:new: :wrench: Add O.S. Inventory maintainer's guide (part of #30)
This commit is a first attempt at creating a maintainer's guide for the
UNICEF Open Source Inventory. The document defines project conventions
as well as provides resources to learn more about the stack used in the
Inventory project.

This contributes to issue #30 but does not close it. This page is still
missing a diagram/chart of the Continuous Integration pipeline, and how
new updates are deployed from GitHub to the World Wide Web. Once the
chart is added, #30 may be closed as complete.

All original commit messages squashed below for reference.

===

meta: Add structure for Inventory maintainer's guide (part of #30)

This is a first pass at a structure of what goes into the maintainer's
guide, as listed out in issue #30.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

meta: Add label conventions for issues and pull requests (part of #30)

This commit defines five categories of label conventions used for the
Open Source Inventory. It provides a framework for how to organize
labels and what the non-standard labels indicate or mean.

Part of #30, does not close.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

meta: Explain commit message convention and best practices (part of #30)

More explanation about what the commit message conventions are. I'm not
totally happy with how this is explained, but it works for now.

Part of #30, but does not close it.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

meta: Document updating git submodule for the Inventory (part of #30)

This commit explains how I typically update the git submodule whenever
there are new upstream changes to pull into the submodule. This way of
organization makes it easier for us to create an independently-useful
theme, while still meeting the immediate need of publishing UNICEF Open
Source best practices.

Part of #30 but does not close.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

meta: Explain how I prioritize development decisions (part of #30)

This lists and orders the different stakeholders of the UNICEF Open
Source Inventory. It adds perspective of who is involved "behind the
scenes" in the future of this project, and who it was primarily designed
for.

It is not exactly governance, but it is better than nothing. It is a
start until this project needs to grow and scale with something more
specific and defined.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

meta: Add AsciiDoc cheat sheet and syntax resources (part of #30)

This commit copies the AsciiDoc cheat sheet I wrote once for Fedora
Linux and put it into this resource. Best part is, CC BY-SA 4.0 licensed
content! We are all compatible.

Part of #30 but does not close it.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

meta: Add Hugo documentation and resources (part of #30)

This commit adds more information about Hugo and links to some helpful
resources to learn more. I also created a reading list of suggested docs
pages to get started with learning about Hugo.

This should make it easier for someone to become more familiar with the
stack and understand how the UNICEF Open Source Inventory works.

Part of #30, but does not close it.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>