:bar_chart: meta: Add diagram of CI pipeline workflow (closes #30)
This commit adds a diagram of the current CI pipeline workflow into the
maintainer's guide. This provides clarity over how changes flow from
source on GitHub to HTML in production.

There is room for improvement, but I believe this is sufficient to get
things rolling and to help train up new contributors. This completes the
first edition of the maintainer's guide.

Closes #30.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>