:microphone: hardware, reading list: Add HBR attribution to blockquote
This commit leverages some of the advanced AsciiDoc syntax to more
properly attribute the blockquote text and to be clear that it is NOT
licensed under the default license of the website (CC-BY-SA 4.0). Given
that the site makes a disclaimer that all content is licensed under the
CC-BY-SA 4.0, extra care needs to be given when "re-printing" content
that is not licensed as such and is shared/published under different
parts of United States copyright law.

See §107 of the Copyright Law of the United States (Title 17) for
additional context on how this excerpt is published in the UNICEF Open
Source Inventory.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>