:new: :mexico: l10n: Add base Spanish translations for Inventory (#50)
* :new: :mexico: l10n: Add base Spanish translations for Inventory

This is a first attempt at integrating multilingual content into the
Open Source Inventory. Since the theme has good support for multiple
languages, it is helpful for us to try and leverage this.

For now, I am translating Spanish since I know the language, but there
is possibility to add others in the future too.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

* :broom: config: Clean up comment headers, put like pieces together

One-off clean-up of the Hugo config file to make it easier for humans to
parse and read. :robot: Addresses @Idadelveloper's review feedback in
PR #50.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>