:beetle: :wrench: static: Show site favicon in browser tab/window
This commit fixes the site favicon. It was always in place, but the
theme is hard-coded to search for it in the path `images/favicon.png`.

Since it is easier to make the change here rather than in our downstream
fork of the theme, I am making the adjustment site-wide for all images
and content that is served from the `static/` directory.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>