:package: :repeat: themes: Synchronize git submodule (part of #33) (#57)
This commit updates the theme submodule to include an option to download
any page from the Open Source Inventory as a PDF. This was added by
@Idadelveloper in unicef/inventory-hugo-theme#9.

Furthermore, we plan to upstream this patch in the near future per the
discussion in themefisher/dot-hugo-documentation-theme#167. Therefore,
this commit contributes to #33 but will not close the issue. We will
close #33 once the patch is merged upstream.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>