Add asciidoc config for automatic TOC (#58)
* config: Add asciidoc config for automatic TOC

This fixes issue #29. Be default hugo turns off the asciidoc automatic TOC since it has its own TOC.
Setting preserveTOC to 'true' instead of 'false' turns on asciidoc's TOC.

Signed-off-by: Ida Delphine <mida@unicef.org>

* :scissors: Remove settings only specifying default options

This commit removes the settings that reflect the default configuration
of how the asciidocExt settings are set. We only need to change the ones
that we need as non-defaults. While this is less explicit, I believe it
decreases complexity in the interest of readability.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

* :page_with_curl: content: Standardize use of automatic TOC when used

`:toc:` must be specified at the top of the document under the front-
matter to render, like other document settings. This commit does the
following when the case applies:

1. For pages with misplaced `:toc:` - set each on new line
2. For pages with manual TOC: Deleted, replaced with `:toc:`

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

Co-authored-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>