:copyright: config: Add attribution to ThemeFisher for upstream theme
This commit makes an additional clarifying note in the copyright
information in the footer of the website. It notes that the site theme
shown on the Inventory is maintained by UNICEF, and also notes our
upstream relationship to the ThemeFisher Dot theme, which greatly
accelerated our work in launching our knowledgebase. :raised_hands:

Follow-up to the request in themefisher/dot-hugo-documentation-theme#167
raised by upstream. CC: @Idadelveloper @@mehedi-sharif @tfsomrat

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>