:new: :memo: missions: Create a template for writing new Missions
This commit adds a draft page of a template that can be reused for
creating new Missions in the future. It also adds documentation as a new
Meta page to describe writing new Missions for other content authors.

This is contributing work to #51. I realized that Missions were not
written in a common format, and this disrupts the user experience. Using
a common template and documenting it also reduces bus factor to include
other writers in the process.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>