:new: :compass: missions: Add a Project Charter Mission (closes #51) (#65)
* :new: :compass: missions: Add a Project Charter Mission (closes #51)

This commit adds a new Mission for Project Charters. It provides a
blueprint for what goes into defining a charter in an Open Source
context. This resource is drafted in support of the incoming August 2021
cohort in the UNICEF Innovation Fund.

Closes #51.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

* :pencil: Second edit pass.

This commit is a second pass with edits to the Project Charter Mission.
Most notably, it adds several examples into the document for the vision,
mission, and community statements.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

* :beetle: :link: Add an incomplete hyperlink

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

* :pen: Add edit to address feedback from @GeorgLink

Distinguishing an internal Open Source policy from a project charter.

https://github.com/unicef/inventory/pull/65#issuecomment-904599190

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

* :pen: Address feedback from @germonprez

Reframe Code of Conduct section a bit and change "stakeholder" to
"project members" in most cases.

https://github.com/unicef/inventory/pull/65#issuecomment-904732851

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

* :pen: Address feedback from @ElizabethN

Better clarify the role of transparency in the project by creating a
charter. This is mostly a riff from the same words in the Pull Request
because it was framed so nicely!

https://github.com/unicef/inventory/pull/65#issuecomment-904766608

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

* :pen: Address feedback from @funnelfiasco

Fix a typo and add a new section to briefly expound on the importance of
building a revision process into a project charter.

https://github.com/unicef/inventory/pull/65#pullrequestreview-740822212

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

* :heart: Add special thanks to volunteer reviews

Special thanks to @GeorgLink, @germonprez, @ElizabethN, and
@funnelfiasco for their time and review on this document!

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>