:book: :wrench: meta: Create new Inventory sub-section
This commit creates a new sub-section for the Open Source Inventory site
underneath the Meta category. This follows on to the change introduced
in #70 to make it easier to navigate content in this category, as the
amount of content continues to grow.

It also gives us more design context for #61.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>