:new: :fist_right: :fist_left: meta: Add On-boarding guide (closes #36) (#72)
This commit adds an on-boarding guide, or the Terms of Reference, for
new inductees to the UNICEF Open Source Mentorship programme. It is
adapted from the Terms of Reference document, originally drafted in a
Google Doc. This takes that Terms of Reference document, converts it to
AsciiDoc, and better integrates it into existing Inventory content.

https://docs.google.com/document/d/1U_YnIx7Oe1D5O5sBN9FmI2GaC8D93P972v8jcs63EHQ/edit?usp=sharing

There is more content to add here, but for now, this is a good first
step. The Google Doc will be wiped out and replaced with a link out to
this page once it is published.

Closes #36.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>