:new: :tm: legal-policy: Add trademark strategy overview (closes #49)
This commit adds a new page to the "Legal & Policy" category titled,
"Trademark strategies". It includes an introduction to what trademarks
are, why they matter in an Open Source context, and it gives a specific
case study of the Moodle learning management system as a model for
trademark management and business model integration.

The purpose of this page is to provide guidance and nonbinding advice to
UNICEF Venture Fund companies in defining their own trademarks and
crafting a strategy around the management of project marks. This
compliments other content on defining project charters and governance.

Closes #49.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>