:speaking_head: :currency_exchange: i18n: Add localization support
This commit introduces the `i18n/` directory as exemplified from the
upstream theme. The only noticeable change is that it adds a new title
to the top of the category section on the index page (`Explore the`
`Inventory by subject`). Adding this here provides greater opportunity
to introduce localization to the site over time.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>