:alembic: ci: Add verbosity to git commands, use env var for repo name
This commit improves the deploy script used in the CI pipeline to make
it easier to use for debugging. `git` commands have added verbosity to
be displayed in the CircleCI build logs. I also changed the hard-coded
repository name to an environment variable already injected into the
container runtime by CircleCI. This makes it easier to re-purpose for
other CircleCI projects.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>