:judge: :telescope: legal-policy: Add Creative Commons primer
This commit creates a new page that introduces the Creative Commons
license family in a UNICEF context. It introduces the licenses, explains
the difference between our preferred Free Cultural licenses and other
Creative Commons license variants, and gives the reader clarification on
how to learn more about the licenses.

This can serve as a common, public reference point for the licenses we
permit, and maybe more importantly, why.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>