legal-policy: Add CC0 as both an acceptable and Free Culture license
Addresses feedback from @lacabra in PR #84 about the Creative Commons
Zero license, which is a Creative Commons license compatible with the
Free Cultural Works definition. It is also one that is acceptable for
the Fund, but I forgot to include it because the license page is buried
on the creativecommons.org web space compared to the other licenses.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>