:judge: :telescope: legal-policy: Add Creative Commons primer (#84)
* :judge: :telescope: legal-policy: Add Creative Commons primer

This commit creates a new page that introduces the Creative Commons
license family in a UNICEF context. It introduces the licenses, explains
the difference between our preferred Free Cultural licenses and other
Creative Commons license variants, and gives the reader clarification on
how to learn more about the licenses.

This can serve as a common, public reference point for the licenses we
permit, and maybe more importantly, why.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

* legal-policy: Add CC0 as both an acceptable and Free Culture license

Addresses feedback from @lacabra in PR #84 about the Creative Commons
Zero license, which is a Creative Commons license compatible with the
Free Cultural Works definition. It is also one that is acceptable for
the Fund, but I forgot to include it because the license page is buried
on the creativecommons.org web space compared to the other licenses.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>