:alembic: :beetle: ci: Do not fail deploy if $CIRCLE_USERNAME is not set
This commit fixes a bug in the deploy script identified by CircleCI
a support agent. The `$CIRCLE_USERNAME` variable is only set if a user
has logged into CircleCI previously.

This fix will cause the CI pipeline to get the `$CIRCLE_USERNAME` and
use it if it is set, but otherwise it will use a fixed string if the
committer has not logged in. Any committer should be able to get a
successful CI pipeline from here on.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>