:recycle: :weight_lifting: content: Remove weight front matter (#102)
This commit removes the `weight` front matter from all content, except
for posts where it is actually necessary. This allows the existing
content to be sorted by default in alphabetical order, while some
content can be "promoted" to a custom order on an as-needed basis.

The purpose of this commit is reduce unnecessary front matter and make
drafting new content more straightforward. Custom orders can be used in
categories for specific pages when needed, like the Open Hardware intro
and overview of the Open Source Mentorship program.

Props to @Idadelveloper for digging into this one and verifying the
behavior. When I tested this in 2019, the mixed alphabetical and custom
order sort did not work.

Closes unicef/inventory-hugo-theme#14.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>