:package: :repeat: theme: Pull upstream changes from theme
This commit re-syncs the Inventory theme with latest changes in the
upstream repo. New commits were rebased into the Inventory theme from
the Dot theme repo, so the git history was reshuffled again.

This in advance of new changes being introduced in the theme as part of
the next couple of sprints.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>