:memo: :boom: Remove "Design" category in favor of SustainOSS Design (#109)
This commit removes the Design category from the Inventory and the site
menu. The design category remained one of the slimmest categories and
lacked content. In light of the work with the SustainOSS Design & UX
Working Group, they are launching a new site using the UNICEF Inventory
theme to create a knowledge-base of design topics and areas of interest.
The knowledge on design in open source will live there.

Closes unicef/inventory#83.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>