:memo: :arrow_up: Code of Conduct: Upgrade to CC v2.1, update contact
This commit upgrades the Code of Conduct, taken from the Contributor
Covenant, from version 2.0 to 2.1. This includes minor changes to the
community statement around race and color.

I also updated contact information to distinguish the contact point as
for the UNICEF Office of Global Innovation and not UNICEF globally. I
also edited "Innovation Fund" to "Venture Fund" per upcoming changes to
the fund's branding.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>