:new: :memo: meta: Add intern on-boarding guide from @Idadelveloper
This commit adds an intern on-boarding guide to help future interns get
up to speed with the Open Source Inventory and how things are set up.
@Idadelveloper wrote this in the final week of her internship, and we
will continue to use this as the preferred way to bring on new
contributors who might work on the project for longer than 1-2 months.

ref: https://docs.google.com/document/d/1eDh2QLJ-Iwz_ELGn9oIj4MvtqC1iMIFPbGro0K52zVE/edit?usp=sharing

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>