:boom: Remove extraneous files
Archetypes are handled in the theme and should be inherited, not defined
here in the downstream site. The `syntax.css` file was added when the
project was launching and later learned about the Hugo-provided
`highlight` function for code blocks.

This continues an effort to clean up the root directory to simplify
maintenance of the static sites.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>