Housekeeping (complements unicef/inventory-hugo-theme#109) (#133)
This commit consolidates several minor changes and repository clean-up that complements the consolidation done by @Neha9849 in unicef/inventory-hugo-theme#109. See commit messages for details:

---

* :boom: Remove all layouts (moved via unicef/inventory-hugo-theme#109)

This commit drops the layouts from the Open Source Inventory site. This
way, they will not be maintained from this repository and future changes
will have to go upstream instead.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

---

* :globe_with_meridians: Use theme-provided localized strings

Drop the internationalization configuration and instead use the
translated strings in the upstream. The only localized content that
should live here is for written content (i.e. files).

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

---

* :boom: Remove extraneous files

Archetypes are handled in the theme and should be inherited, not defined
here in the downstream site. The `syntax.css` file was added when the
project was launching and later learned about the Hugo-provided
`highlight` function for code blocks.

This continues an effort to clean up the root directory to simplify
maintenance of the static sites.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

---

* :package: :arrow_up: theme: Update to latest upstream

This commit syncs the theme git submodule with the upstream repo.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

---

* :wrench: config: Add editPrefix option for edit button

This adds support for the "Edit this Page" button, so it correctly
points to the online repository for the source content.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>