:building_construction: Outreachy 2022: DPG Standard restructure
This commit represents the sum changes for migrating the UNICEF Open
Source Inventory to a new support structure modeled around the Digital
Public Good Standard. Based on user testing research, the interface is
made to be more useful for a project or Venture Fund start-up company
aiming to achieve compliance with the Digital Public Good Standard.

Redirects with page aliases are in place so all URL addresses from the
old organizational model will redirect to the correct place.

Closes unicef/inventory-hugo-theme#126.

===

:memo: Change DPG Indicators page type to dpg-standard

This adopts the expected layout for the DPG Indicators, with a mini
profile for each indicator.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>