:camera_flash: static: Use a more thematic banner image on home page
This commit changes the banner image used on the site home page to be
something more thematic to UNICEF's core work and also represent the
overall theme of learning and education. The goal is to improve
messaging of what the site is for. Photo by Aaron Burden on Unsplash.

Fixes unicef/inventory-hugo-theme#153.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>