:wrench: toolkits: Move Toolkits from navbar to a new category
This changes the user experience to make the user more aware that by
visiting another UNICEF Toolkit, they will go to a different website.
This addresses feedback from user research testing that found the
"Toolkits" navbar option to be confusing, since the website and content
changed but the general UI/UX remains the same.

This change makes the redirect more obvious. It leverages the page
redirect used previously in the Data Transparency & Ethics category in
the Open Source Inventory.

Closes unicef/inventory-hugo-theme#156.