:building_construction: Outreachy 2022: DPG Standard restructure (#136)
This commit represents the sum changes for migrating the UNICEF Open
Source Inventory to a new support structure modeled around the Digital
Public Good Standard. Based on user testing research, the interface is
made to be more useful for a project or Venture Fund start-up company
aiming to achieve compliance with the Digital Public Good Standard.

Redirects with page aliases are in place so all URL addresses from the
old organizational model will redirect to the correct place.

Closes unicef/inventory-hugo-theme#126.

===

:memo: Change DPG Indicators page type to dpg-standard

This adopts the expected layout for the DPG Indicators, with a mini
profile for each indicator.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

===

* :wrench: toolkits: Move Toolkits from navbar to a new category

This changes the user experience to make the user more aware that by
visiting another UNICEF Toolkit, they will go to a different website.
This addresses feedback from user research testing that found the
"Toolkits" navbar option to be confusing, since the website and content
changed but the general UI/UX remains the same.

This change makes the redirect more obvious. It leverages the page
redirect used previously in the Data Transparency & Ethics category in
the Open Source Inventory.

Closes unicef/inventory-hugo-theme#156

===

* :memo: DPG-1: Use a standard post instead of redirect type

A page redirect doesn't work on index pages. Instead, this page should
be populated with basic information and the reader is encouraged to
visit the SDG goals website for more information.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

===

* :memo: Finalize content, fix redirects, and prepare for deployment

This commit includes several minor to moderate changes that prepares the
DPG Standard indicator redesign to deploy to production:

1. Add `keywords` to post front-matter in order to better utilize the
   search bar function.
2. Remove `categories` front-matter from posts since this metadata is
   implicitly inherited by path.
3. Add and/or modify `tags` in many posts.
4. Fix aliases missing a starting `/` so they work as expected
   (previously, many aliases were incomplete and not working).
5. Minor formatting and syntax: Remove unnecessary quotemarks, change
   order of some front-matter, use alphabetical sorts, etc.
6. Reorganize some articles from one DPG Indicator to another. This
   happened mostly among DPG Indicators 2, 3, and 8.
7. Use Themify icon set for home page categories. Only set an `icon`
   front-matter for content categories that appear on the site index
   page.
8. Propose pages for deletion, using a `warning` admonition to give
   ample notice that a page may disappear from the site.
9. Utilize new site features in older content, like the page author and
   table of contents.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>
Co-authored-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>