:memo: faq: Add questions (closes #35)
This commit reworks the Frequently Asked Questions (F.A.Q.) page on the
site to include more useful information about the UNICEF Open Source
Inventory, how it is maintained, who is responsible for it, how to
contribute, and more.

This also changes the file from an AsciiDoc to Markdown file because the
content inside of the FAQ shortcodes must be Markdown in order to
render. Therefore, it does not make sense to keep the file as AsciiDoc
when all the contents of the file are written in Markdown.

Closes unicef/inventory#35.
Closes unicef/inventory-hugo-theme#162.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>